fix - write component에서 사용하는 함수들 변경 * utlls.js 를 utils.js로 변경 * utils.js를 사용하도록 변경.
