+++ asset/component/ListItem.js
@@ -2,7 +2,7 @@ import $ from '../util/domControll';
 import ElementBuilder from './ElementBuilder';
 import IconBtns from './IconButtons';
 import Image from './Image';
-import { stringEllipsis } from '../util/utlls';
+import { stringEllipsis } from '../util/utils';
 import '../css/listItem.css';
 
 function Comment(comment) {
